Move transformer model into test_padded_tensor
